requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
